Endpoints book-detail and create-reservation solved

+++ src/main/kotlin/ar/edu/unsam/phm/repository/CrudBookRepository.kt
@@ -3,9 +3,11 @@ package ar.edu.unsam.phm.repository
 import ar.edu.unsam.phm.domain.Book
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.CrudRepository
+import org.springframework.data.repository.query.Param
+import java.time.LocalDate
 import java.util.Optional
 
-interface CrudBookRepository: CrudRepository<Book, Int> {
+interface CrudBookRepository: CrudRepository<Book, Long> {
     fun findByIsbn(isbn: String): Optional<Book>
 
     @Query("""
